add new config for tiflash
