+++ pkg/manager/member/tiflash_util.go
@@ -21,6 +21,7 @@ import (
 
 	"github.com/pingcap/tidb-operator/pkg/apis/pingcap/v1alpha1"
 	"github.com/pingcap/tidb-operator/pkg/controller"
+	"github.com/pingcap/tidb-operator/pkg/util/cmpver"
 
 	corev1 "k8s.io/api/core/v1"
 )
@@ -29,6 +30,11 @@ const (
 	defaultClusterLog = "/data0/logs/flash_cluster_manager.log"
 	defaultErrorLog   = "/data0/logs/error.log"
 	defaultServerLog  = "/data0/logs/server.log"
+)
+
+var (
+	// the first version that tiflash change default config
+	tiflashEqualOrGreaterThanV540, _ = cmpver.NewConstraint(cmpver.GreaterOrEqual, "v5.4.0")
 )
 
 func buildTiFlashSidecarContainers(tc *v1alpha1.TidbCluster) ([]corev1.Container, error) {
@@ -98,6 +104,114 @@ func buildSidecarContainer(name, path, image string,
 			fmt.Sprintf("touch %s; tail -n0 -F %s;", path, path),
 		},
 	}
+}
+
+func GetTiFlashConfig(tc *v1alpha1.TidbCluster) *v1alpha1.TiFlashConfigWraper {
+	version := tc.TiFlashVersion()
+	if ok, err := tiflashEqualOrGreaterThanV540.Check(version); err == nil && ok {
+		return getTiFlashConfigV2(tc)
+	}
+	return getTiFlashConfig(tc)
+}
+
+func getTiFlashConfigV2(tc *v1alpha1.TidbCluster) *v1alpha1.TiFlashConfigWraper {
+	config := tc.Spec.TiFlash.Config.DeepCopy()
+	if config == nil {
+		config = v1alpha1.NewTiFlashConfig()
+	}
+	if config.Common == nil {
+		config.Common = v1alpha1.NewTiFlashCommonConfig()
+	}
+	if config.Proxy == nil {
+		config.Proxy = v1alpha1.NewTiFlashProxyConfig()
+	}
+	common := config.Common
+	proxy := config.Proxy
+
+	name := tc.Name
+	ns := tc.Namespace
+	clusterDomain := tc.Spec.ClusterDomain
+	ref := tc.Spec.Cluster.DeepCopy()
+	noLocalPD := tc.HeterogeneousWithoutLocalPD()
+	noLocalTiDB := ref != nil && ref.Name != "" && tc.Spec.TiDB == nil
+
+	// common
+	{
+		// storage
+		if common.Get("path") == nil && common.Get("storage") == nil {
+			paths := []string{}
+			for i := range tc.Spec.TiFlash.StorageClaims {
+				paths = append(paths, fmt.Sprintf("/data%d/db", i))
+			}
+			if len(paths) == 0 {
+				paths = []string{"/data0/db"}
+			}
+			common.Set("storage.main.dir", paths)
+		}
+		if common.Get("raft.kvstore_path") == nil {
+			common.SetIfNil("storage.raft.dir", "/data0/kvstore")
+		}
+
+		// port
+		common.SetIfNil("tcp_port", int64(9000))
+		common.SetIfNil("http_port", int64(8123))
+
+		// flash
+		tidbStatusAddr := fmt.Sprintf("%s.%s.svc:10080", controller.TiDBMemberName(name), ns)
+		if noLocalTiDB {
+			tidbStatusAddr = fmt.Sprintf("%s.%s.svc%s:10080",
+				controller.TiDBMemberName(ref.Name), ref.Namespace, controller.FormatClusterDomain(ref.ClusterDomain))
+		}
+		common.SetIfNil("flash.tidb_status_addr", tidbStatusAddr)
+		common.SetIfNil("flash.service_addr", "0.0.0.0:3930")
+		common.SetIfNil("flash.flash_cluster.log", defaultClusterLog)
+		common.SetIfNil("flash.proxy.addr", "0.0.0.0:20170")
+		common.SetIfNil("flash.proxy.advertise-addr", fmt.Sprintf("%s-POD_NUM.%s.%s.svc%s:20170", controller.TiFlashMemberName(name), controller.TiFlashPeerMemberName(name), ns, controller.FormatClusterDomain(clusterDomain)))
+		common.SetIfNil("flash.proxy.data-dir", "/data0/proxy")
+		common.SetIfNil("flash.proxy.config", "/data0/proxy.toml")
+
+		// logger
+		common.SetIfNil("logger.errorlog", defaultErrorLog)
+		common.SetIfNil("logger.log", defaultServerLog)
+
+		// raft
+		pdAddr := fmt.Sprintf("%s.%s.svc:2379", controller.PDMemberName(name), ns)
+		if len(clusterDomain) > 0 {
+			pdAddr = "PD_ADDR"
+		} else if noLocalPD {
+			pdAddr = fmt.Sprintf("%s.%s.svc%s:2379", controller.PDMemberName(ref.Name), ref.Namespace, controller.FormatClusterDomain(ref.ClusterDomain))
+		}
+		common.SetIfNil("raft.pd_addr", pdAddr)
+	}
+
+	// proxy
+	{
+		proxy.SetIfNil("log-level", "info")
+		proxy.SetIfNil("server.engine-addr", fmt.Sprintf("%s-POD_NUM.%s.%s.svc%s:3930", controller.TiFlashMemberName(name), controller.TiFlashPeerMemberName(name), ns, controller.FormatClusterDomain(clusterDomain)))
+		proxy.SetIfNil("server.status-addr", "0.0.0.0:20292")
+		proxy.SetIfNil("server.advertise-status-addr", fmt.Sprintf("%s-POD_NUM.%s.%s.svc%s:20292", controller.TiFlashMemberName(name), controller.TiFlashPeerMemberName(name), ns, controller.FormatClusterDomain(clusterDomain)))
+	}
+
+	// Note the config of tiflash use "_" by convention, others(proxy) use "-".
+	if tc.IsTLSClusterEnabled() {
+		common.Set("security.ca_path", path.Join(tiflashCertPath, corev1.ServiceAccountRootCAKey))
+		common.Set("security.cert_path", path.Join(tiflashCertPath, corev1.TLSCertKey))
+		common.Set("security.key_path", path.Join(tiflashCertPath, corev1.TLSPrivateKeyKey))
+		common.SetIfNil("tcp_port_secure", int64(9000))
+		common.SetIfNil("https_port", int64(8123))
+		common.Del("http_port")
+		common.Del("tcp_port")
+
+		proxy.Set("security.ca-path", path.Join(tiflashCertPath, corev1.ServiceAccountRootCAKey))
+		proxy.Set("security.cert-path", path.Join(tiflashCertPath, corev1.TLSCertKey))
+		proxy.Set("security.key-path", path.Join(tiflashCertPath, corev1.TLSPrivateKeyKey))
+
+		if commonVal, proxyVal := common.Get("security.cert_allowed_cn"), proxy.Get("security.cert-allowed-cn"); commonVal != nil && proxyVal == nil {
+			proxy.Set("security.cert-allowed-cn", commonVal.Interface())
+		}
+	}
+
+	return config
 }
 
 func getTiFlashConfig(tc *v1alpha1.TidbCluster) *v1alpha1.TiFlashConfigWraper {
